+++ src/test/java/io/anserini/encoder/sparse/SpladePlusPlusEnsembleDistilEncoderInference2Test.java
@@ -14,11 +14,9 @@
  * limitations under the License.
  */
 
-package io.anserini.encoder;
+package io.anserini.encoder.sparse;
 
 import ai.onnxruntime.OrtException;
-import io.anserini.encoder.sparse.SparseEncoder;
-import io.anserini.encoder.sparse.SpladePlusPlusEnsembleDistilEncoder;
 
 import org.junit.Test;
